Migrated php5 tests
