+++ test/snapshot/php5.test.js
@@ -1,0 +1,14 @@
+var parser = require("../../src/index");
+
+describe('Test syntax parsing without PHP7 support', function() {
+
+  it('special keywords should fail', function() {
+    var ast = parser.parseEval('class foo { function list() { } }', {
+      parser: {
+        php7: false,
+        suppressErrors: true
+      }
+    });
+    expect(ast).toMatchSnapshot();
+  });
+});
